fix(examples): updated fastify example to show multi middleware example
