+++ examples/nestjs-fastify/src/user.schema.ts
@@ -1,17 +1,5 @@
 import { z } from 'zod';
 
-export function createPaginatedResponseSchema<ItemType extends z.ZodTypeAny>(
-  itemSchema: ItemType,
-) {
-  return z.object({
-    pageIndex: z.number(),
-    pageSize: z.number(),
-    totalCount: z.number(),
-    totalPages: z.number(),
-    items: z.array(itemSchema),
-  });
-}
-
 export const userSchema = z.object({
   name: z.string(),
   email: z.string(),
